only run condition check on highlight cell
