+++ packages/web-frontend/src/components/View/MatrixWrapper/components/RowGroup.js
@@ -9,7 +9,6 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import DropDownArrowIcon from 'material-ui/svg-icons/navigation/arrow-drop-down';
 import shallowEqual from 'shallowequal';
-import { getPresentationOption } from '../../../../utils';
 
 import Cell from './Cell';
 
@@ -92,7 +91,6 @@ export default class RowGroup extends Component {
           {columns.slice(startColumn, displayedColumnCount).map((column, index) => {
             const isCellActive = index === highlightedColumn && isRowHighlighted;
             const value = columnData ? columnData[column.key] : '';
-            const presentation = getPresentationOption(presentationOptions, value);
 
             return (
               <div
@@ -103,8 +101,8 @@ export default class RowGroup extends Component {
                   cellKey={index}
                   onMouseEnter={() => onCellMouseEnter(index, rowId)}
                   onMouseLeave={() => onCellMouseLeave()}
-                  onClick={() => onCellClick(presentation, value)}
-                  color={presentation ? presentation.color : { color: '' }}
+                  onCellClick={onCellClick}
+                  presentationOptions={presentationOptions}
                   value={value}
                   columnActiveStripStyle={styles.columnActiveStrip}
                   isActive={isCellActive}
